% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{AsyncHookOutput}
\alias{AsyncHookOutput}
\title{Create an AsyncHookOutput}
\usage{
AsyncHookOutput(async_timeout = NULL)
}
\arguments{
\item{async_timeout}{Integer or NULL. Timeout in milliseconds.}
}
\value{
Object of class \code{AsyncHookOutput}.
}
\description{
Signals that the hook will complete asynchronously.
}
\examples{
out <- AsyncHookOutput(async_timeout = 5000L)
out$asyncTimeout
}
