% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protocol.R
\name{build_control_error_response}
\alias{build_control_error_response}
\title{Build an error control-response JSON string}
\usage{
build_control_error_response(request_id, error_msg)
}
\description{
Build an error control-response JSON string
}
\keyword{internal}
