% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{ForkSessionResult}
\alias{ForkSessionResult}
\title{Result of a fork_session() operation}
\usage{
ForkSessionResult(session_id)
}
\arguments{
\item{session_id}{Character. UUID of the new forked session.}
}
\value{
Object of class \code{ForkSessionResult}.
}
\description{
Result of a fork_session() operation
}
\examples{
res <- ForkSessionResult(session_id = "aaaaaaaa-bbbb-cccc-dddd-eeeeeeeeeeee")
res$session_id
}
