% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{McpServerInfo}
\alias{McpServerInfo}
\title{Create a McpServerInfo}
\usage{
McpServerInfo(name, version)
}
\arguments{
\item{name}{Character.}

\item{version}{Character.}
}
\value{
Object of class \code{McpServerInfo}.
}
\description{
Create a McpServerInfo
}
