% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{McpStatusResponse}
\alias{McpStatusResponse}
\title{Create a McpStatusResponse}
\usage{
McpStatusResponse(mcp_servers)
}
\arguments{
\item{mcp_servers}{List of \code{McpServerStatus}.}
}
\value{
Object of class \code{McpStatusResponse}.
}
\description{
Create a McpStatusResponse
}
