% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{McpToolInfo}
\alias{McpToolInfo}
\title{Create a McpToolInfo}
\usage{
McpToolInfo(name, description = NULL, annotations = NULL)
}
\arguments{
\item{name}{Character.}

\item{description}{Character or NULL.}

\item{annotations}{List or NULL.}
}
\value{
Object of class \code{McpToolInfo}.
}
\description{
Create a McpToolInfo
}
