% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{NotificationHookSpecificOutput}
\alias{NotificationHookSpecificOutput}
\title{Hook-specific output for Notification hook}
\usage{
NotificationHookSpecificOutput(additional_context = NULL)
}
\arguments{
\item{additional_context}{Character or NULL.}
}
\value{
Object of class \code{NotificationHookSpecificOutput}.
}
\description{
Hook-specific output for Notification hook
}
