% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{PermissionRequestHookSpecificOutput}
\alias{PermissionRequestHookSpecificOutput}
\title{Hook-specific output for PermissionRequest hook}
\usage{
PermissionRequestHookSpecificOutput(decision)
}
\arguments{
\item{decision}{List. Required. Permission decision payload.}
}
\value{
Object of class \code{PermissionRequestHookSpecificOutput}.
}
\description{
Hook-specific output for PermissionRequest hook
}
