% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{PostToolUseFailureHookInput}
\alias{PostToolUseFailureHookInput}
\title{Create a PostToolUseFailureHookInput}
\usage{
PostToolUseFailureHookInput(
  session_id,
  transcript_path,
  cwd,
  tool_name,
  tool_input,
  tool_use_id,
  error,
  is_interrupt = NULL,
  permission_mode = NULL,
  agent_id = NULL,
  agent_type = NULL
)
}
\arguments{
\item{session_id}{Character.}

\item{transcript_path}{Character.}

\item{cwd}{Character.}

\item{tool_name}{Character.}

\item{tool_input}{List.}

\item{tool_use_id}{Character.}

\item{error}{Character.}

\item{is_interrupt}{Logical or NULL.}

\item{permission_mode}{Character or NULL.}

\item{agent_id}{Character or NULL.}

\item{agent_type}{Character or NULL.}
}
\value{
Object of class \code{PostToolUseFailureHookInput}.
}
\description{
Create a PostToolUseFailureHookInput
}
