% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{PostToolUseFailureHookSpecificOutput}
\alias{PostToolUseFailureHookSpecificOutput}
\title{Hook-specific output for PostToolUseFailure hook}
\usage{
PostToolUseFailureHookSpecificOutput(additional_context = NULL)
}
\arguments{
\item{additional_context}{Character or NULL.}
}
\value{
Object of class \code{PostToolUseFailureHookSpecificOutput}.
}
\description{
Hook-specific output for PostToolUseFailure hook
}
