% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{PreToolUseHookSpecificOutput}
\alias{PreToolUseHookSpecificOutput}
\title{Hook-specific output for PreToolUse hook}
\usage{
PreToolUseHookSpecificOutput(
  permission_decision = NULL,
  permission_decision_reason = NULL,
  updated_input = NULL,
  additional_context = NULL
)
}
\arguments{
\item{permission_decision}{Character or NULL. One of \code{"allow"}, \code{"deny"}, \code{"ask"}.}

\item{permission_decision_reason}{Character or NULL.}

\item{updated_input}{List or NULL. Modified tool input.}

\item{additional_context}{Character or NULL.}
}
\value{
Object of class \code{PreToolUseHookSpecificOutput}.
}
\description{
Hook-specific output for PreToolUse hook
}
