% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\docType{data}
\name{RATE_LIMIT_TYPES}
\alias{RATE_LIMIT_TYPES}
\title{Valid rate-limit type values (\code{RateLimitType})}
\format{
An object of class \code{character} of length 5.
}
\usage{
RATE_LIMIT_TYPES
}
\description{
Which rate-limit window applies.
}
\keyword{datasets}
