% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{SandboxIgnoreViolations}
\alias{SandboxIgnoreViolations}
\title{Create a SandboxIgnoreViolations}
\usage{
SandboxIgnoreViolations(file = NULL, network = NULL)
}
\arguments{
\item{file}{Character vector or NULL.}

\item{network}{Character vector or NULL.}
}
\value{
Object of class \code{SandboxIgnoreViolations}.
}
\description{
Create a SandboxIgnoreViolations
}
