% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{SandboxSettings}
\alias{SandboxSettings}
\title{Create SandboxSettings}
\usage{
SandboxSettings(
  enabled = NULL,
  auto_allow_bash_if_sandboxed = NULL,
  excluded_commands = NULL,
  allow_unsandboxed_commands = NULL,
  network = NULL,
  ignore_violations = NULL,
  enable_weaker_nested_sandbox = NULL
)
}
\arguments{
\item{enabled}{Logical or NULL.}

\item{auto_allow_bash_if_sandboxed}{Logical or NULL.}

\item{excluded_commands}{Character vector or NULL.}

\item{allow_unsandboxed_commands}{Logical or NULL.}

\item{network}{\code{SandboxNetworkConfig} or NULL.}

\item{ignore_violations}{\code{SandboxIgnoreViolations} or NULL.}

\item{enable_weaker_nested_sandbox}{Logical or NULL.}
}
\value{
Object of class \code{SandboxSettings}.
}
\description{
Create SandboxSettings
}
\examples{
sb <- SandboxSettings(enabled = TRUE, allow_unsandboxed_commands = FALSE)
sb$enabled
}
