% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\docType{data}
\name{SERVER_TOOL_NAMES}
\alias{SERVER_TOOL_NAMES}
\title{Valid server-side tool names (\code{ServerToolName})}
\format{
An object of class \code{character} of length 8.
}
\usage{
SERVER_TOOL_NAMES
}
\description{
The set of tools the API executes server-side on the model's behalf. Branch
on \code{ServerToolUseBlock$name} to know which server tool was invoked.
}
\keyword{datasets}
