% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{SessionStartHookSpecificOutput}
\alias{SessionStartHookSpecificOutput}
\title{Hook-specific output for SessionStart hook}
\usage{
SessionStartHookSpecificOutput(additional_context = NULL)
}
\arguments{
\item{additional_context}{Character or NULL.}
}
\value{
Object of class \code{SessionStartHookSpecificOutput}.
}
\description{
Hook-specific output for SessionStart hook
}
