% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{StopHookInput}
\alias{StopHookInput}
\title{Create a StopHookInput}
\usage{
StopHookInput(
  session_id,
  transcript_path,
  cwd,
  stop_hook_active,
  permission_mode = NULL
)
}
\arguments{
\item{session_id}{Character.}

\item{transcript_path}{Character.}

\item{cwd}{Character.}

\item{stop_hook_active}{Logical.}

\item{permission_mode}{Character or NULL.}
}
\value{
Object of class \code{StopHookInput}.
}
\description{
Create a StopHookInput
}
