% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{StreamEvent}
\alias{StreamEvent}
\title{Create a StreamEvent}
\usage{
StreamEvent(uuid, session_id, event, parent_tool_use_id = NULL)
}
\arguments{
\item{uuid}{Character.}

\item{session_id}{Character.}

\item{event}{List. Raw Anthropic API stream event.}

\item{parent_tool_use_id}{Character or NULL.}
}
\value{
Object of class \code{StreamEvent}.
}
\description{
Create a StreamEvent
}
