% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{SubagentStartHookInput}
\alias{SubagentStartHookInput}
\title{Create a SubagentStartHookInput}
\usage{
SubagentStartHookInput(
  session_id,
  transcript_path,
  cwd,
  agent_id,
  agent_type,
  permission_mode = NULL
)
}
\arguments{
\item{session_id}{Character.}

\item{transcript_path}{Character.}

\item{cwd}{Character.}

\item{agent_id}{Character.}

\item{agent_type}{Character.}

\item{permission_mode}{Character or NULL.}
}
\value{
Object of class \code{SubagentStartHookInput}.
}
\description{
Create a SubagentStartHookInput
}
