% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{SubagentStartHookSpecificOutput}
\alias{SubagentStartHookSpecificOutput}
\title{Hook-specific output for SubagentStart hook}
\usage{
SubagentStartHookSpecificOutput(additional_context = NULL)
}
\arguments{
\item{additional_context}{Character or NULL.}
}
\value{
Object of class \code{SubagentStartHookSpecificOutput}.
}
\description{
Hook-specific output for SubagentStart hook
}
