% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{SubagentStopHookInput}
\alias{SubagentStopHookInput}
\title{Create a SubagentStopHookInput}
\usage{
SubagentStopHookInput(
  session_id,
  transcript_path,
  cwd,
  stop_hook_active,
  agent_id,
  agent_transcript_path,
  agent_type,
  permission_mode = NULL
)
}
\arguments{
\item{session_id}{Character.}

\item{transcript_path}{Character.}

\item{cwd}{Character.}

\item{stop_hook_active}{Logical.}

\item{agent_id}{Character.}

\item{agent_transcript_path}{Character.}

\item{agent_type}{Character.}

\item{permission_mode}{Character or NULL.}
}
\value{
Object of class \code{SubagentStopHookInput}.
}
\description{
Create a SubagentStopHookInput
}
