% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{SystemMessage}
\alias{SystemMessage}
\title{Create a SystemMessage}
\usage{
SystemMessage(subtype, data)
}
\arguments{
\item{subtype}{Character. Subtype string.}

\item{data}{List. Raw data dict.}
}
\value{
Object of class \code{SystemMessage}.
}
\description{
Create a SystemMessage
}
