% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{SystemPromptPreset}
\alias{SystemPromptPreset}
\title{Create a SystemPromptPreset}
\usage{
SystemPromptPreset(exclude_dynamic_sections = NULL, append = NULL)
}
\arguments{
\item{exclude_dynamic_sections}{Logical or NULL.}

\item{append}{Character or NULL. Additional instructions to append.}
}
\value{
Object of class \code{SystemPromptPreset}.
}
\description{
Create a SystemPromptPreset
}
\examples{
sp <- SystemPromptPreset(exclude_dynamic_sections = TRUE)
sp$type  # "preset"
}
