% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{TERMINAL_TASK_STATUSES}
\alias{TERMINAL_TASK_STATUSES}
\title{Terminal task statuses (a task in one of these will not transition further)}
\usage{
TERMINAL_TASK_STATUSES
}
\description{
Terminal task statuses (a task in one of these will not transition further)
}
