% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{ThinkingConfigDisabled}
\alias{ThinkingConfigDisabled}
\title{Create a ThinkingConfigDisabled}
\usage{
ThinkingConfigDisabled()
}
\value{
Object of class \code{ThinkingConfigDisabled}.
}
\description{
Create a ThinkingConfigDisabled
}
