% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{UserPromptSubmitHookInput}
\alias{UserPromptSubmitHookInput}
\title{Create a UserPromptSubmitHookInput}
\usage{
UserPromptSubmitHookInput(
  session_id,
  transcript_path,
  cwd,
  prompt,
  permission_mode = NULL
)
}
\arguments{
\item{session_id}{Character.}

\item{transcript_path}{Character.}

\item{cwd}{Character.}

\item{prompt}{Character.}

\item{permission_mode}{Character or NULL.}
}
\value{
Object of class \code{UserPromptSubmitHookInput}.
}
\description{
Create a UserPromptSubmitHookInput
}
