% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{UserPromptSubmitHookSpecificOutput}
\alias{UserPromptSubmitHookSpecificOutput}
\title{Hook-specific output for UserPromptSubmit hook}
\usage{
UserPromptSubmitHookSpecificOutput(additional_context = NULL)
}
\arguments{
\item{additional_context}{Character or NULL.}
}
\value{
Object of class \code{UserPromptSubmitHookSpecificOutput}.
}
\description{
Hook-specific output for UserPromptSubmit hook
}
