% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{ContextUsageResponse}
\alias{ContextUsageResponse}
\title{Create a ContextUsageResponse}
\usage{
ContextUsageResponse(
  categories,
  total_tokens,
  max_tokens = NULL,
  raw_max_tokens = NULL,
  percentage = NULL,
  model = NULL,
  is_auto_compact_enabled = NULL,
  memory_files = NULL,
  mcp_tools = NULL,
  agents = NULL,
  grid_rows = NULL,
  auto_compact_threshold = NULL,
  deferred_builtin_tools = NULL,
  system_tools = NULL,
  system_prompt_sections = NULL,
  slash_commands = NULL,
  skills = NULL,
  message_breakdown = NULL,
  api_usage = NULL
)
}
\arguments{
\item{categories}{List of \code{ContextUsageCategory}.}

\item{total_tokens}{Integer.}
}
\value{
Object of class \code{ContextUsageResponse}.
}
\description{
Create a ContextUsageResponse
}
\examples{
cats <- list(ContextUsageCategory("user", 1024L, "#4e79a7"))
resp <- ContextUsageResponse(cats, total_tokens = 1024L)
resp$totalTokens
}
