% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{ContextUsageResponse}
\alias{ContextUsageResponse}
\title{Create a ContextUsageResponse}
\usage{
ContextUsageResponse(categories, total_tokens)
}
\arguments{
\item{categories}{List of \code{ContextUsageCategory}.}

\item{total_tokens}{Integer.}
}
\value{
Object of class \code{ContextUsageResponse}.
}
\description{
Create a ContextUsageResponse
}
\examples{
cats <- list(ContextUsageCategory("user", 1024L, "#4e79a7"))
resp <- ContextUsageResponse(cats, total_tokens = 1024L)
resp$totalTokens
}
