% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{PermissionResultAllow}
\alias{PermissionResultAllow}
\title{Allow permission result}
\usage{
PermissionResultAllow(updated_input = NULL, updated_permissions = NULL)
}
\arguments{
\item{updated_input}{List or NULL. Modified tool input.}

\item{updated_permissions}{List or NULL. Permission updates.}
}
\value{
Object of class \code{PermissionResultAllow}.
}
\description{
Allow permission result
}
