% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{RateLimitEvent}
\alias{RateLimitEvent}
\title{Create a RateLimitEvent}
\usage{
RateLimitEvent(rate_limit_info, uuid, session_id)
}
\arguments{
\item{rate_limit_info}{A \code{RateLimitInfo} object.}

\item{uuid}{Character.}

\item{session_id}{Character.}
}
\value{
Object of class \code{RateLimitEvent}.
}
\description{
Create a RateLimitEvent
}
