% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{ResultMessage}
\alias{ResultMessage}
\title{Create a ResultMessage}
\usage{
ResultMessage(
  subtype,
  duration_ms,
  duration_api_ms,
  is_error,
  num_turns,
  session_id,
  stop_reason = NULL,
  total_cost_usd = NULL,
  usage = NULL,
  result = NULL,
  structured_output = NULL,
  model_usage = NULL,
  permission_denials = NULL,
  errors = NULL,
  uuid = NULL
)
}
\arguments{
\item{subtype}{Character.}

\item{duration_ms}{Integer.}

\item{duration_api_ms}{Integer.}

\item{is_error}{Logical.}

\item{num_turns}{Integer.}

\item{session_id}{Character.}

\item{stop_reason}{Character or NULL.}

\item{total_cost_usd}{Numeric or NULL.}

\item{usage}{List or NULL.}

\item{result}{Character or NULL.}

\item{structured_output}{Any or NULL.}

\item{model_usage}{List or NULL.}

\item{permission_denials}{List or NULL.}

\item{errors}{List or NULL.}

\item{uuid}{Character or NULL.}
}
\value{
Object of class \code{ResultMessage}.
}
\description{
Create a ResultMessage
}
