% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{SandboxNetworkConfig}
\alias{SandboxNetworkConfig}
\title{Create a SandboxNetworkConfig}
\usage{
SandboxNetworkConfig(
  allow_unix_sockets = NULL,
  allow_all_unix_sockets = NULL,
  allow_local_binding = NULL,
  http_proxy_port = NULL,
  socks_proxy_port = NULL
)
}
\arguments{
\item{allow_unix_sockets}{Character vector or NULL.}

\item{allow_all_unix_sockets}{Logical or NULL.}

\item{allow_local_binding}{Logical or NULL.}

\item{http_proxy_port}{Integer or NULL.}

\item{socks_proxy_port}{Integer or NULL.}
}
\value{
Object of class \code{SandboxNetworkConfig}.
}
\description{
Create a SandboxNetworkConfig
}
