% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{TaskUsage}
\alias{TaskUsage}
\title{Create a TaskUsage}
\usage{
TaskUsage(total_tokens, tool_uses)
}
\arguments{
\item{total_tokens}{Integer.}

\item{tool_uses}{Integer.}
}
\value{
Object of class \code{TaskUsage}.
}
\description{
Create a TaskUsage
}
