% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\docType{data}
\name{TERMINAL_TASK_STATUSES}
\alias{TERMINAL_TASK_STATUSES}
\title{Terminal task statuses (a task in one of these will not transition further)}
\format{
An object of class \code{character} of length 3.
}
\usage{
TERMINAL_TASK_STATUSES
}
\description{
Terminal task statuses (a task in one of these will not transition further)
}
\keyword{datasets}
