% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{ThinkingConfigAdaptive}
\alias{ThinkingConfigAdaptive}
\title{Create a ThinkingConfigAdaptive}
\usage{
ThinkingConfigAdaptive()
}
\value{
Object of class \code{ThinkingConfigAdaptive}.
}
\description{
Create a ThinkingConfigAdaptive
}
