% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{ThinkingConfigEnabled}
\alias{ThinkingConfigEnabled}
\title{Create a ThinkingConfigEnabled}
\usage{
ThinkingConfigEnabled(budget_tokens)
}
\arguments{
\item{budget_tokens}{Integer. Token budget for thinking.}
}
\value{
Object of class \code{ThinkingConfigEnabled}.
}
\description{
Create a ThinkingConfigEnabled
}
\examples{
cfg <- ThinkingConfigEnabled(budget_tokens = 5000L)
cfg$budget_tokens
}
